ADD: added game and user deletion, untested

+++ PlayerHive/src/main/java/com/unipi/PlayerHive/service/AdminService.java
@@ -2,34 +2,45 @@ package com.unipi.PlayerHive.service;
 
 import com.unipi.PlayerHive.DTO.games.AddGameDTO;
 import com.unipi.PlayerHive.DTO.games.EditGameDTO;
+import com.unipi.PlayerHive.DTO.users.GameOwnerDTO;
 import com.unipi.PlayerHive.config.Exceptions.ResourceAlreadyExistsException;
 import com.unipi.PlayerHive.model.Game;
 import com.unipi.PlayerHive.model.GameNeo4j;
+import com.unipi.PlayerHive.repository.ReviewRepository;
 import com.unipi.PlayerHive.repository.games.GameNeo4jRepository;
 import com.unipi.PlayerHive.repository.games.GameRepository;
-import com.unipi.PlayerHive.utility.GameMapper;
+import com.unipi.PlayerHive.utility.batch.UserConsistencyManager;
+import com.unipi.PlayerHive.utility.map.GameMapper;
+import jakarta.annotation.Nonnull;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
+import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.ArrayList;
 import java.util.NoSuchElementException;
+import java.util.stream.Stream;
 
 @Service
 public class AdminService {
     private final GameRepository gameRepository;
     private final GameNeo4jRepository gameNeo4jRepository;
     private final GameMapper gameMapper;
+    private final MongoTemplate mongoTemplate;
 
-    public AdminService(GameRepository gameRepository, GameNeo4jRepository gameNeo4jRepository, GameMapper gameMapper) {
+    private final ReviewRepository reviewRepository;
+
+    public AdminService(GameRepository gameRepository, GameNeo4jRepository gameNeo4jRepository, GameMapper gameMapper, MongoTemplate mongoTemplate, ReviewRepository reviewRepository) {
         this.gameRepository = gameRepository;
         this.gameNeo4jRepository = gameNeo4jRepository;
         this.gameMapper = gameMapper;
+        this.mongoTemplate = mongoTemplate;
+        this.reviewRepository = reviewRepository;
     }
 
     @Transactional
-    public void addGame(@Valid @RequestBody AddGameDTO newGame) {
+    public void addGame(@Nonnull @Valid @RequestBody AddGameDTO newGame) {
 
         if(gameRepository.existsByName(newGame.getName()))
                 throw new ResourceAlreadyExistsException("Game "+ newGame.getName() +" already exists");
@@ -51,7 +62,7 @@ public class AdminService {
     }
 
     @Transactional
-    public void editGame(String gameId, @Valid @RequestBody EditGameDTO editGame) {
+    public void editGame(String gameId, @Nonnull @Valid @RequestBody EditGameDTO editGame) {
 
         Game game = gameRepository.findById(gameId).orElseThrow(() -> new NoSuchElementException("The Game with id:\"" + gameId + "\" does not exist"));
         // is there a better way??
@@ -108,8 +119,36 @@ public class AdminService {
     }
 
     @Transactional
-    public void deleteGame(String gameId) { // fix all user libraries <TODO>
-        gameRepository.deleteById(gameId);
+    public void deleteGame(String gameId) {
+
+        if(!gameRepository.existsById(gameId)){
+            throw new NoSuchElementException("The game chosen for deletion does not exist");
+        }
+
+        // we obtain from neo4j all the relationships that point to the game of interest
+        Stream<GameOwnerDTO> allOwners = gameNeo4jRepository.findGameOwnersOf(gameId);
+
+        //the user consistency manager requires mongoTemplate for batch operation
+        UserConsistencyManager userManager = new UserConsistencyManager(mongoTemplate);
+
+        //all users "hoursPlayed" and "numGames" are decreased accordingly
+        userManager.adjustUserStatsAfterGameRemoval(allOwners.iterator());
+
+        allOwners.close();
+
+        //all reviews are now deleted
+        reviewRepository.removeByGameId(gameId);
+
+
+
+        //the game node in neo4j is removed
         gameNeo4jRepository.deleteById(gameId);
+
+        // all reviews of the game are removed from the reviews array present in every user document
+        userManager.removeGameReviewsFromUsers(gameId,gameRepository);
+
+        //we can finally delete the JSON document
+        gameRepository.deleteById(gameId);
     }
+
 }
